% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car_color.R
\docType{data}
\name{car_color_palettes}
\alias{car_color_palettes}
\title{Palette of Car Colors}
\format{
An object of class \code{list} of length 3.
}
\usage{
car_color_palettes
}
\arguments{
\item{name}{Name of the color palette. \code{datsun1} has 7 distinct colors from the Datsun motor company. \code{datsun2} has 5 distinct colors and \code{datsun3} has four colors.}

\item{n}{The amount of colors desired for your plot.}
}
\value{
A vector of colors.
}
\description{
This function includes color palettes of car companies that previous decades. More color palettes will be added in the future.
}
\examples{

car_color("datsun1", 4)
car_color("datsun2")

library(tidyverse)
mtcars \%>\%
ggplot(aes(hp, mpg)) +
geom_point(aes(color = as.factor(cyl))) +
scale_color_manual(values = car_color("datsun1"))
}
\keyword{datasets}
