% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_bar.R
\name{freq_bar}
\alias{freq_bar}
\title{Function to get proportions and percentages of a single variable}
\usage{
freq_bar(data, x, fill, total)
}
\arguments{
\item{data}{The data frame that includes the factor variable you are interested in getting proportions, percentages, and a visual of groups in the variable.}

\item{x}{The X variable you'd like to examine.}

\item{fill}{Value to determine the color you'd like your bar graph to be filled with. The outline of the histograms is set to "White"}

\item{total}{Numerical value of your total sample in your study.}
}
\value{
Returns a list with a visual of the groups and a table of the proportions and percentages
}
\description{
This function is to examine normality and linearity of a linear regression in the form of histograms of your predictor and your outcome as well as a scatterplot of these variables.
}
\examples{

To examine the visual
freq_bar(mtcars, as.factor(cyl), fill = "dodgerblue", total = 32)[[1]]

To create a table
freq_bar(mtcars, as.factor(cyl), fill = "dodgerblue", total = 32)[[2]]
}
