% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_residuals.R
\name{linear_residuals}
\alias{linear_residuals}
\title{Assessment of Residuals in Linear Regression}
\usage{
linear_residuals(
  data,
  x,
  y,
  se = c(TRUE, FALSE),
  alpha,
  size,
  line_size,
  color,
  line_color
)
}
\arguments{
\item{data}{The data frame that includes the variables you are interested in examining in a linear regression.}

\item{x}{Predictors to be included in your linear regression. You can either include one variable in quotations (e.g., "hp" from the mtcars dataset)
or you can create an object of predictors (e.g., predictors <- c("disp", "cyl", "wt")).}

\item{y}{Your outcome of interest in your linear regression. This variable does not require quotations.}

\item{se}{A logical vector to decide if you'd like to include the standard error of your linear regression.}

\item{alpha}{Value to determine how transparent you'd like your points in this function's plots.}

\item{size}{Value to determine the size of the points in this function's plots. You can also determine if you'd like to assign these values to a
categorical variable in your dataset.}

\item{line_size}{Value to decide if you'd like your lines to be thinner or thicker in this function's plots.}

\item{color}{Value to determine what color you'd like your points to be in this function's plots (e.g., "blue", "#6a1f25")}

\item{line_color}{value to determine what color you'd like your lines to be in this function's plots (e.g., "blue", "#6a1f25")}
}
\value{
Returns two ggplot2 visuals showing the residuals in your model.
}
\description{
This function is to examine residuals to assess for homoscedasticity and normality in a linear regression.
This function allows for the examination of residuals of a bivariate linear regression or a multiple linear regression.
}
\examples{

One Predictor
residual_view(data = mtcars, x = "hp", y = mpg, se = FALSE, alpha = .3, size = 2, line_size = 1, color = "blue", line_color = "dodgerblue")

predictors <- c("hp", "carb", "gear")
residual_view(data = mtcars, x = predictors, y = mpg, se = FALSE, alpha = .3, size = 2, line_size = 1, color = "blue", line_color = "dodgerblue")
}
