% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_bi_dist.R
\name{uni_bi_dist}
\alias{uni_bi_dist}
\title{Assessment of Normality and Linearity in Linear Regression}
\usage{
uni_bi_dist(
  data,
  x,
  y,
  alpha,
  bins,
  fill,
  color,
  loess_color,
  line_color,
  se = c(TRUE, FALSE),
  size
)
}
\arguments{
\item{data}{The data frame that includes the variables you are interested in examining in a linear regression.}

\item{x}{The X variable you'd like to examine.}

\item{y}{Your outcome of interest.}

\item{alpha}{Value to determine how transparent you'd like your histograms and points in the scatterplot.}

\item{bins}{Value to adjust the bins of your histogram.}

\item{fill}{Value to determine the color you'd like your histogram to be filled with. The outline of the histograms is set to "White"}

\item{color}{Value to determine what color you'd like your points to be in the scatterplot (e.g., "blue", "#6a1f25")}

\item{loess_color}{value to determine what color you'd like your loess line to be in the scatterplot (e.g., "blue", "#6a1f25")}

\item{line_color}{value to determine what color you'd like your linear relationship to be in the scatterplot (e.g., "blue", "#6a1f25")}

\item{se}{A logical vector to decide if you'd like to include the standard error of both the loess and linear relationships in your scatterplot.}

\item{size}{Value to decide if you'd like your lines to be thinner or thicker in your scatterplot}
}
\value{
Returns three ggplot2 visuals. Two histograms of your variables of interest and a scatterplot of the relationship.
}
\description{
This function is to examine normality and linearity of a linear regression in the form of histograms of your predictor and your outcome as well as a scatterplot of these variables.
}
\examples{

uni_bi_dist(data = mtcars, x = hp, y = mpg, alpha = .8, bins = 15, fill = "dodgerblue", color = "black", loess_color = "darkgreen", line_color = "green", se = FALSE, size = 1.25)
}
